+++ src/vue/modules/requests/sponsorship-requests/components/requests-table.vue
@@ -129,7 +129,7 @@ export default {
 
 .requests-table__review-btn {
   font-size: 1.2rem;
-  color: $col-primary-lighten;
+  color: $col-btn;
   cursor: pointer;
 }
 </style>