+++ src/vue/pages/Downloads.vue
@@ -20,7 +20,12 @@
           {{ 'downloads-page.mobile-apps-header' | globalize }}
         </h3>
         <p class="downloads__section-text">
-          {{ 'downloads-page.mobile-apps-desc' | globalize }}
+          {{ 'downloads-page.mobile-apps-desc' | globalize(
+            {
+              appName: config.APP_NAME,
+            }
+          )
+          }}
         </p>
 
         <div class="downloads__phones-wrapper">